requirements.txt
numpy
PyYaml
pylint
paho-mqtt<2.0.0